+++ drawio-custom-plugins/src/index.ts
@@ -1,2 +1,3 @@
 import "./linkSelectedNodeWithData";
 import "./liveshare";
+import "./focus";