+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.0-alpha.1]
+## [0.8.0-alpha.2]
 
 ### Added
 
 -   Adds a status bar item to quickly change the current drawio theme.
--   Adds drawio-language-mode (see [#130](https://github.com/hediet/vscode-drawio/issues/130))
+-   Adds drawio-language-mode (see [#130](https://github.com/hediet/vscode-drawio/issues/130)).
+-   Users of the Insiders Build are asked for feedback.
 
 ### Changed